Added permissions to the dev commands

+++ commands/development/restart.js
@@ -4,14 +4,16 @@ module.exports = {
     name: "restart",
     description: "Restarts the bot",
     execute(client, message, args, Discord) {
-        message.channel.send("Restarting...").then(msg => {
-            client.destroy()
-            client.login(process.env.TOKEN)
-            msg.edit("Restart successful!")
-            client.user.setPresence({
-                activity: { name: `BWTinder Bot is helping people find love in ${client.guilds.cache.size} servers! | *help`},
-                status: "dnd"
+        if (message.member.roles.cache.find(role => role.name === "Developer")) {
+            message.channel.send("Restarting...").then(msg => {
+                client.destroy()
+                client.login(process.env.TOKEN)
+                msg.edit("Restart successful!")
+                client.user.setPresence({
+                    activity: { name: `BWTinder Bot is helping people find love in ${client.guilds.cache.size} servers! | *help`},
+                    status: "dnd"
+                })
             })
-        })
+        }
     }
 }
