+++ commands/development/shutdown.js
@@ -2,8 +2,10 @@ module.exports = {
     name: "shutdown",
     description: "Shuts down the bot",
     execute(client, message, args, Discord) {
-        message.channel.send("Shutting down...").then(m => {
-            client.destroy()
-        })
+        if (message.member.roles.cache.find(role => role.name === "Developer")) {
+            message.channel.send("Shutting down...").then(m => {
+                client.destroy()
+            })
+        }
     }
 }
